Change Integer to double

+++ src/main/java/app/entities/Order.java
@@ -35,7 +35,7 @@ public class Order {
         this.timestamp = timestamp;
     }
 
-    public Order(String title, int carportWidth, int carportLength, OrderStatus status, Integer totalPrice, List<OrderBillItem> orderBill, LocalDateTime timestamp) {
+    public Order(String title, int carportWidth, int carportLength, OrderStatus status, double totalPrice, List<OrderBillItem> orderBill, LocalDateTime timestamp) {
         this.title = title;
         this.carportWidth = carportWidth;
         this.carportLength = carportLength;
